simple-app/requirements.txt
fastapi==0.104.0
psycopg2-binary
pydantic==2.4.2
uvicorn
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
requests==2.31.0
gunicorn